+++ projects/framework-houdini/release_notes.md
@@ -1,5 +1,11 @@
 # ftrack Framework Houdini integration release Notes
 
+
+# Upcoming
+
+* [changed] Host, Client instance; Pass run_in_main_thread argument.
+
+
 ## v24.6.0
 2024-06-26